Changing datatypes to library standards

+++ srcs/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: tlouekar <tlouekar@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/11 13:37:41 by tlouekar          #+#    #+#             */
-/*   Updated: 2020/07/01 12:26:52 by tlouekar         ###   ########.fr       */
+/*   Updated: 2020/07/06 10:24:38 by tlouekar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,6 +16,7 @@
 # include "../libft/includes/libft.h"
 # include <fcntl.h>
 # include <stdarg.h>
+# include <stdint.h>
 
 static char			g_flags[20] = "0123456789#-+hlL.* ";
 
@@ -64,7 +65,7 @@ char			*parse(char *str, t_printf *data);
 char			*preparse(char *str, va_list args, t_printf *data);
 char			*structurize(t_printf *data, char *ptr);
 t_printf		*initialize(char *ptr);
-char			*helper_itoa_base(long long value, int base, t_printf *data);
+char			*helper_itoa_base(intmax_t value, int base, t_printf *data);
 char			*ftoa(long double value, t_printf *data);
 void			helper_print_padding(t_printf *data);
 void			helper_plusminus(t_printf *data);
